@@ -14,7 +14,7 @@ description: SakhaTyla app
 version: 2.0.7+207
 
 environment:
-  sdk: ">=2.5.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
