@@ -3,22 +3,17 @@ import 'dart:async';
 
 class Debouncer {
   final int milliseconds;
-  VoidCallback action;
-  Timer _timer;
+  Timer? _timer;
 
-  Debouncer({this.milliseconds});
+  Debouncer({required this.milliseconds});
 
   run(VoidCallback action) {
-    if (_timer != null) {
-      _timer.cancel();
-    }
+    _timer?.cancel();
 
     _timer = Timer(Duration(milliseconds: milliseconds), action);
   }
 
   cancel() {
-    if (_timer != null) {
-      _timer.cancel();
-    }
+    _timer?.cancel();
   }
 }
